Remove verta.repository (keeping Blob)
